Fix build on i386
On old systems int is limited to range 2147483647 : -2147483648

+++ NumToWords/EntryAr.go
@@ -38,7 +38,7 @@ func (Entry *EntryAr) Translate(Input int) string {
 		{
 			Result = Entry.handleMillions(Input)
 		}
-	case Input < 1000000000000: //1000000000-9999999999
+	case int64(Input) < 1000000000000: //1000000000-9999999999
 		{
 			Result = Entry.handleBillions(Input)
 		}
